ci: separate is_indexed and extensive vector search tests
